AddAuthProvider, Remove eslintrc.cjs, loginPage footer
