--- .eslintrc.cjs
@@ -1,28 +1,0 @@
-module.exports = {
-  env: {
-    browser: true,
-    es2021: true,
-  },
-  extends: [
-    'plugin:react/recommended',
-    'airbnb',
-    'airbnb-typescript',
-    'prettier',
-  ],
-  overrides: [],
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-    project: ['./tsconfig.json'],
-    tsconfigRootDir: '.',
-    createDefaultProgram: true,
-  },
-  plugins: ['react'],
-  rules: {
-    'react/jsx-filename-extension': [1, { extensions: ['.ts', '.tsx'] }],
-    'import/no-extraneous-dependencies': [
-      'error',
-      { devDependencies: ['vite.config.ts'] },
-    ],
-  },
-};
