+++ src/containers/auth/login/LoginContainer.tsx
@@ -1,8 +1,6 @@
 import LoginForm from '@components/features/auth/login/forms/LoginForms'
 import LoginFormTitle from '@components/features/auth/login/LoginFormTitle'
-import { AccountCircle, Key } from '@mui/icons-material'
-import { Button, Container, Grid, InputAdornment, Link, TextField, Typography } from '@mui/material'
-import { grey } from '@mui/material/colors'
+import { Grid } from '@mui/material'
 import React from 'react'
 
 function LoginContainer() {
@@ -13,8 +11,7 @@ function LoginContainer() {
         direction='column'
         justifyContent='center'
         alignItems='center'
-        // TODO: 어느 페이지에 적용해야할까요? 화면에 레이아웃이 스크롤이 없고 간단한..곳들
-        sx={{ height: 'calc( 100dvh - 77px)' }}
+        sx={{ height: 'calc( 100dvh - 105px)' }}
       >
         <LoginFormTitle />
         <LoginForm />
