+++ src/pages/layout/baseLayout/footer/BaseLayoutFooter.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 const BaseLayoutFooter = () => {
   return (
-    <div className='mt-2 text-center text-sm text-gray-400'>
+    <div className='my-4 text-center text-sm text-gray-400'>
       Copyright © Neulsang Corp. All rights reserved.
     </div>
   )
